move utils.dbfields to fields.encrypted
